+++ packages/f-checkout/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.17.0
+------------------------------
+*December 7, 2020*
+
+### Added
+- Component tests for 'Collection' prop variant.
+
+### Changed
+- Split component tests into separate spec files
+
 
 v0.16.1
 ------------------------------